+++ tests/unit/active.account.spec.js
@@ -15,9 +15,6 @@ describe('Accounts.vue', () => {
     store = new Vuex.Store({})
     store.registerModule('accounts', accountsM)
     store.registerModule('drizzle', drizzleM)
-    store.state.drizzle.initialized = true
-    store.state.accounts.activeAccount =
-      '0x048ffCA67A523a4B51EA7a251168FD86e623f243'
   })
 
   it('displays active account if Drizzle is initialized', () => {
@@ -28,7 +25,23 @@ describe('Accounts.vue', () => {
         convertedBalance: () => '100'
       }
     })
+    store.state.drizzle.initialized = true
+    store.state.accounts.activeAccount =
+      '0x048ffCA67A523a4B51EA7a251168FD86e623f243'
     const div = wrapper.find('div')
     expect(div.text()).toContain(store.getters['accounts/activeAccount'])
+  })
+
+  it(`displays 'Loading...' if Drizzle is not initialized`, () => {
+    const wrapper = shallowMount(Accounts, {
+      localVue,
+      store,
+      computed: {
+        convertedBalance: () => '100'
+      }
+    })
+    store.state.drizzle.initialized = false
+    const div = wrapper.find('div')
+    expect(div.text()).toContain('Loading...')
   })
 })
